Request available days of history on start
And limit the seek bar to this number if there are less than
31 days available.

Also insert an initial record if the database is empty. This
can only happen when the user starts the app for the very first
time in which case he/she wants to see that the app is already
counting. At it can only count from this very moment on.