Handle system shutdowns
Because Android doesn't send ACTION_SCREEN_OFF when the device is
shutting down.

Without this commit the app was counting the down time as screen time
because there would be a SCREEN_ON and an no SCREEN_OFF until the device
comes back online.

Unfortunately, the ACTION_SHUTDOWN broadcast is not always and everywhere
sent. For example, the Pixel 2 seems to never send it, while the Pixel 3a
(with the latest Android) does.