Save on/off times in one record
This architecture makes it easier to update the off time of the
ongoing record (whenever the notification is updated).

This way we are no longer dependent on ACTION_SCREEN_OFF, which
may not come under certain circumstances (like when the device is
suddenly out of power).

This also greatly simplifies querying the database and requires
less code. Should have done it this way from the start.