Add POST_NOTIFICATIONS permission for Android > 13
Because Android 13 (API level 33) requires apps to request the new
POST_NOTIFICATIONS permission to post notifications.

Since this app needs to run a Foreground Service to not get killed
by the system, and because running a Foreground Service requires an
on going notification, this app needs to request the new permission.