Add menu option to ignore battery optimization
Since there are still a lot of devices with a really aggressive
battery optimization.

Of course, the foreground service alone should be enough for the
system to keep this app running. That's what a foreground service
is there for after all. But apparently that's not enough for some
device manufacturers.

So this adds an easy option to disable battery optimization from
within the app.